feat throwing exceptions in case of api failure

+++ src/controllers/FileController.ts
@@ -9,38 +9,54 @@ class FileController {
   public async store(request: Request, response: Response): Promise<Response> {
     const { file } = request;
 
-    const fileCreated = await File.create({
-      name: file?.originalname,
-      size: file?.size,
-      key: file?.key,
-      url: file?.location,
-    });
+    if (!file) throw new Error('missing file');
 
-    return response.json(fileCreated);
+    try {
+      const fileCreated = await File.create({
+        name: file?.originalname,
+        size: file?.size,
+        key: file?.key,
+        url: file?.location,
+      });
+
+      return response.json(fileCreated);
+    } catch (err) {
+      throw new Error('file creation failed');
+    }
   }
 
   public async index(request: Request, response: Response): Promise<Response> {
-    const files = await File.find();
+    try {
+      const files = await File.find();
 
-    return response.json(files);
+      return response.json(files);
+    } catch (err) {
+      throw new Error('failed to list all files');
+    }
   }
   public async delete(request: Request, response: Response): Promise<Response> {
-    const { file_id } = request.params;
+    try {
+      const { file_id } = request.params;
 
-    const file = await File.findById(file_id);
+      const file = await File.findById(file_id);
 
-    const s3 = new aws.S3();
+      if (!file) throw new Error('file does not exist');
 
-    if (process.env.STORAGE_TYPE === 's3')
-      s3.deleteObject({ Bucket: 'upload-aws', Key: file.key }).promise();
-    else
-      promisify(fs.unlink)(
-        path.resolve(__dirname, '..', '..', 'temp', 'uploads', file.key),
-      );
+      const s3 = new aws.S3();
 
-    file.remove();
+      if (process.env.STORAGE_TYPE === 's3')
+        s3.deleteObject({ Bucket: 'upload-aws', Key: file.key }).promise();
+      else
+        promisify(fs.unlink)(
+          path.resolve(__dirname, '..', '..', 'temp', 'uploads', file.key),
+        );
 
-    return response.json({ message: 'Delete successfully' });
+      file.remove();
+
+      return response.json({ message: 'Delete successfully' });
+    } catch (err) {
+      throw new Error('failed to delete faile');
+    }
   }
 }
